pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "fakturoid-v3"
version = "2.0.0"
dependencies = [
  "pydantic >= 2.0",
  "pydantic[email]",
  "requests==2.32.3",
  "python-dateutil",
  "sybil[pytest]",
  "httpx>=0.28.1",
  "deprecated>=1.2.18",
  "authlib>=1.6.0",
  "pytest-asyncio>=1.0.0",
]
requires-python = ">=3.12"
authors = [
  {name = "Roman Krejcik", email = "farin@farin.cz"},
  {name = "Jaroslav Henner", email = "jaroslav.henner@gmail.com"},
]
maintainers = [
  {name = "Jaroslav Henner", email = "jaroslav.henner@gmail.com"}
]
description = "The Python interface to online accounting service [Fakturoid](http://fakturoid.cz/)."
readme = "README.md"
license = "MIT"
license-files = ["LICEN[CS]E.*"]
keywords = ["fakturoid", "accounting"]
classifiers = [
    "Development Status :: 1 - Planning",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Topic :: Office/Business :: Financial :: Accounting",
]

[project.optional-dependencies]
test = [
  "freezegun >= 1.5.1",
  "black >= 25.1.0",
  "pre-commit >= 4.2.0",
  "pytest-recording >= 0.13.3",
  "pytest-doctestplus==1.4.0"
]
dev = [
  "mypy",
  'types-requests==2.32.0.20250515'
]

[project.urls]
Homepage = "https://github.com/jarovo/python-fakturoid-v3/"
Repository = "https://github.com/jarovo/python-fakturoid-v3/"

[tool.hatch.build.targets.wheel]
packages = ["fakturoid"]
include = [
  "fakturoid/py.typed"
]

[tool.pytest.ini_options]
#addopts = "--doctest-plus --doctest-modules --doctest-glob='*.md'"
